requirements.txt
Click==7.0
Flask==1.1.1
Flask-SSLify==0.1.5
itsdangerous==0.24
Jinja2==2.10.3
MarkupSafe==0.23
pkg-resources==0.0.0
pylibmc==1.6.1
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2016.4
redis==3.4.1
six==1.10.0
Werkzeug==0.16.0
